+++ apps/autotest/addUBD.sh
@@ -26,3 +26,7 @@ mkdir -p './node_modules/@ub-e'
 if [ ! -d "./node_modules/@ub-e/pdfsign" ]; then
   ln -s "$(pwd)/../../../ub-e/packages/pdfsign" "./node_modules/@ub-e/"
 fi
+
+if [ ! -d "./node_modules/@ub-e/ocr" ]; then
+  ln -s "$(pwd)/../../../ub-e/packages/ocr" "./node_modules/@ub-e/"
+fi