+++ apps/autotest/models/TST/public/forms/tst_document-fm.vue
@@ -82,7 +82,11 @@
       />
 
       <u-grid>
-        <u-auto-field attribute-name="person" />
+        <u-auto-field
+          attribute-name="person"
+          allow-dictionary-adding
+          :build-add-dictionary-config="buildAddPersonConfig"
+        />
         <u-auto-field attribute-name="employee" />
       </u-grid>
     </u-form-container>
@@ -161,6 +165,22 @@ module.exports.default = {
         throw e
       }
     },
+
+    buildAddPersonConfig (cfg) {
+      const parsedFIO = _.compact(cfg.query.split(' ')).map(word => _.capitalize(word))
+      if (!parsedFIO[1]) this.$notify.error('Wrong citizen full name, must contain at least 2 words')
+      cfg.props = {}
+      cfg.props.parentContext = {
+        fullFIO: cfg.query,
+        shortFIO: parsedFIO[1] + ' ' + _.compact([parsedFIO[0] && parsedFIO[0][1], parsedFIO[2] && parsedFIO[2][1]])
+          .join('.') + '.',
+        lastName: parsedFIO[1],
+        firstName: parsedFIO[0],
+        middleName: parsedFIO[2] || null
+      }
+      return cfg
+    },
+
     showDialog () {
       // line below is for testing purpose
       // better to use this.dialogInfo('uba_user') inside Vue instance