+++ packages/adminui-vue/utils/Form/validation.js
@@ -1,55 +1,97 @@
 /* global UB */
 
-const _ = require('lodash')
 const Vue = require('vue')
 // vuex required for type checking
 // eslint-disable-next-line no-unused-vars
 const Vuex = require('vuex')
-const { validationMixin } = require('vuelidate')
 const { required } = require('vuelidate/lib/validators/index')
 
 const { mapInstanceFields } = require('./helpers')
 
-module.exports = class Validator {
+const attrCaptionsMixin = {
+  computed: {
+    attributeCaptions () {
+      const { attributeCaptions } = this.$options
+      if (typeof attributeCaptions === 'function') {
+        return attributeCaptions.call(this)
+      }
+      return attributeCaptions
+    }
+  },
+
+  methods: {
+    getCustomFormCaptionByPath (path) {
+      return path.reduce(getPropByKey, this.attributeCaptions)
+    },
+
+    getAttributeCaption (path) {
+      return this.getCustomFormCaptionByPath(path) ?? getEntityAttributeCaption(this.entitySchema, path[0])
+    }
+  },
+
+  attributeCaptions: {}
+}
+
+/**
+ * Abstraction for easy managing of validation of some form. This class uses a `Vue`
+ * instance with configured `Vuelidate` state to provide reactivity of form data.
+ */
+class Validator {
+  constructor (vueInstance) {
+    this._vueInstance = vueInstance
+  }
+
   /**
-   * Create a Vue instance for form data validation.
+   * Create a instance for form data validation based on some options defined as Vue mixin.
    * Default behavior is to check entity schema attributes with `allowNull=true` and `defaultView=true`.
    *
    * `customValidationMixin` can extend default behavior by it own rules.
    *
-   * @param {Vuex} store Store
-   * @param {UBEntity} entitySchema Entity schema
-   * @param {string[]} masterFieldList Field list of master entity
+   * @param {object} params
+   * @param {Vuex} params.store Store
+   * @param {UBEntity} params.entitySchema Entity schema
+   * @param {string[]} params.masterFieldList Field list of master entity
    * @param {Vue} [customValidationMixin={}] Custom validations what extends default
-   * @return {Vue} Vue instance
    */
-  constructor (store, entitySchema, masterFieldList, customValidationMixin = {}) {
-    this._entitySchema = entitySchema
-
-    const requiredFields = entitySchema
+  static initializeWithCustomOptions ({
+    store,
+    entitySchema,
+    masterFieldList,
+    customValidationMixin = {}
+  }) {
+    const requiredAttributesNames = entitySchema
       .filterAttribute(attr => attr.defaultView && !attr.allowNull && masterFieldList.includes(attr.code))
-      .map(a => a.name)
+      .map(attr => attr.name)
 
     const defaultValidationMixin = {
-      computed: mapInstanceFields(entitySchema.getAttributeNames()),
+      computed: {
+        ...mapInstanceFields(entitySchema.getAttributeNames()),
+
+        entitySchema () {
+          return entitySchema
+        }
+      },
 
       validations () {
-        return Object.fromEntries(requiredFields.map(field => [field, { required }]))
+        return Object.fromEntries(requiredAttributesNames.map(attr => [attr, { required }]))
       }
     }
 
-    this._vueInstance = new Vue({
+    const vueInstance = new Vue({
       store,
       mixins: [
-        validationMixin,
+        attrCaptionsMixin,
         defaultValidationMixin,
         customValidationMixin
       ]
     })
+
+    return new Validator(vueInstance)
   }
 
   /**
    * Returns the current state of validation. The method is useful when you have dynamic validation
+   *
    * @returns {object} Vuelidate object
    */
   getValidationState () {
@@ -57,36 +99,92 @@ module.exports = class Validator {
   }
 
   /**
-   * Validates form data with Vuelidate. If validation is failed `UBAbortError` will be thrown
-   * @throws {UB.UBAbortError}
+   * Get caption by attribute name from `attributeCaptions` sections. If it
+   * is not defined, default locale i18n(`${entity}.${attributeName}`) will be returned
+   *
+   * @param {string} attributeName
+   * @returns {string}
    */
-  validateForm () {
-    const { $v } = this._vueInstance
-    $v.$touch()
-    if ($v.$error) {
-      const masterEntityName = this._entitySchema.code
+  getAttributeCaption (attributeName) {
+    const attributePath = attributeName.split('.')
+    const caption = this._vueInstance.getAttributeCaption(attributePath)
+    return UB.i18n(caption)
+  }
 
-      const invalidFields = new Set()
-      for (const { path } of $v.$flattenParams()) {
-        const attrValidation = path.reduce((acc, cur) => acc[cur], $v)
-        if (attrValidation.$invalid) {
-          invalidFields.add(path.join('.'))
+  /**
+   * Get error text for some first failed validation rule of the attribute
+   *
+   * @param {string} attributeName
+   * @returns {string | null}
+   */
+  getErrorForAttribute (attributeName) {
+    const attributePath = attributeName.split('.')
+    const attrValidationState = attributePath.reduce(getPropByKey, this._vueInstance.$v)
+
+    if (!attrValidationState || !attrValidationState.$error) {
+      return null
+    }
+
+    for (const param in attrValidationState.$params) {
+      if (attrValidationState[param] === false) {
+        const ruleParams = attrValidationState.$params[param]
+        const errorText = ruleParams ? ruleParams.$errorText : null
+        if (errorText) {
+          return UB.i18n(errorText)
         }
       }
-
-      const invalidFieldsCaptions = [...invalidFields].map(field => {
-        const formFieldCaption = _.get(this._vueInstance, `${field}:caption`)
-        if (formFieldCaption) {
-          return formFieldCaption
-        }
-        const localeString = `${masterEntityName}.${field}`
-        return UB.i18n(localeString) === localeString ? field : UB.i18n(localeString)
-      })
-      const errMsg = UB.i18n('validationError', invalidFieldsCaptions.join(', '))
-      const err = new UB.UBError(errMsg)
-      UB.showErrorWindow(err)
-      throw new UB.UBAbortError(errMsg)
     }
+
+    return UB.i18n('requiredField')
+  }
+
+  /**
+   * Check if the attribute has the required rule in the configured validation
+   *
+   * @param {string} attributeName
+   * @returns {boolean}
+   */
+  getIsAttributeRequired (attributeName) {
+    const attributePath = attributeName.split('.')
+    const attrValidationState = attributePath.reduce(getPropByKey, this._vueInstance.$v)
+    return attrValidationState && 'required' in attrValidationState
+  }
+
+  /**
+   * Validates form data with the Vuelidate help. If validation is failed `UBAbortError` will be thrown
+   *
+   * @param {object} [params = {}]
+   * @param {boolean} [params.showErrorModal = true] To display error modal if validation is failed
+   * @param {string} [params.errorMsgTemplate = 'validationError'] Error message template for the error modal
+   * @throws {UB.UBAbortError}
+   */
+  validateForm ({
+    showErrorModal = true,
+    errorMsgTemplate = 'validationError'
+  } = {}) {
+    const { $v } = this._vueInstance
+
+    $v.$touch()
+    if (!$v.$error) {
+      return
+    }
+
+    const invalidFieldsCaptions = new Set()
+    for (const { path } of $v.$flattenParams()) {
+      const attrValidation = path.reduce(getPropByKey, $v)
+      if (attrValidation.$error) {
+        // for array validation remove $each.XXX: aaa.bbb.$each.125.ccc.ddd ->  aa.bbb.$each.125.ccc.ddd
+        const attrCaption = this.getAttributeCaption(path.filter((f, idx, arr) => (f !== '$each') && (arr[idx - 1] !== '$each')).join('.'))
+        invalidFieldsCaptions.add(attrCaption)
+      }
+    }
+    const errMsg = UB.i18n(errorMsgTemplate, [...invalidFieldsCaptions].join(', '))
+
+    if (showErrorModal) {
+      UB.showErrorWindow(errMsg)
+    }
+
+    throw new UB.UBAbortError(errMsg)
   }
 
   /**
@@ -95,4 +193,58 @@ module.exports = class Validator {
   reset () {
     this._vueInstance.$v.$reset()
   }
+}
+
+/**
+ * Helper function that returns an object property by the key. This method is useful for
+ * passing to the `reduce()` array function to get some object nested property
+ *
+ * @param {object|null} obj
+ * @param {string} key
+ * @returns {object|null}
+ */
+function getPropByKey (obj, key) {
+  return obj ? obj[key] : null
+}
+
+/**
+ * Get entity attribute caption in the current locale
+ *
+ * @param {UBEntity} entitySchema
+ * @param {string} attr
+ * @returns {string}
+ */
+function getEntityAttributeCaption (entitySchema, attr) {
+  const localeString = `${entitySchema.code}.${attr}`
+  return UB.i18n(localeString) === localeString ? attr : UB.i18n(localeString)
+}
+
+const validationMixin = {
+  inject: {
+    entitySchema: {
+      from: 'entitySchema',
+      default: {}
+    }
+  },
+
+  mixins: [
+    attrCaptionsMixin
+  ],
+
+  provide () {
+    return {
+      validator: this.validator
+    }
+  },
+
+  computed: {
+    validator () {
+      return new Validator(this)
+    }
+  }
+}
+
+module.exports = {
+  Validator,
+  validationMixin
 }