+++ packages/ub/_hookMetadataTransformation.js
@@ -1,7 +1,7 @@
 module.exports = ubMixinTransformation
 
 /**
- * Adds This hook is called by server in the single thread initialization mode. In this stage
+ * This hook is called by server in the single thread initialization mode. In this stage
  *  - native Domain is not created yet
  *  - js files form models is not evaluated
  * Hook can mutate a Domain JSON, for example - adds additional attributes to the entities and lang files, etc
@@ -15,6 +15,7 @@ function ubMixinTransformation (domainJson, serverConfig) {
   if (mt && (mt.enabled === true)) {
     addMultitenancyMixinAttributes(domainJson, serverConfig)
   }
+  validateAttributesBlobStore(domainJson, serverConfig)
 }
 
 /**
@@ -22,7 +23,7 @@ function ubMixinTransformation (domainJson, serverConfig) {
  * @param {object<string, {modelName: string, meta: object, lang: object<string, object>}>} domainJson
  * @param {object} serverConfig
  */
-function addImplicitlyAddedMixins(domainJson, serverConfig) {
+function addImplicitlyAddedMixins (domainJson, serverConfig) {
   const impl = serverConfig.application.domain.implicitlyAddedMixins
   if (!impl.length) return
   console.debug('Adding implicitlyAddedMixins:', serverConfig.application.domain.implicitlyAddedMixins)
@@ -42,7 +43,7 @@ function addImplicitlyAddedMixins(domainJson, serverConfig) {
  * @param {object<string, {modelName: string, meta: object, lang: object<string, object>}>} domainJson
  * @param {object} serverConfig
  */
-function addMultitenancyMixinAttributes(domainJson, serverConfig) {
+function addMultitenancyMixinAttributes (domainJson, serverConfig) {
   const dbCfg = serverConfig.application.connections
   const connCfgMap = {}
   let defaultConn
@@ -62,9 +63,9 @@ function addMultitenancyMixinAttributes(domainJson, serverConfig) {
         dbDefault = "(COALESCE(current_setting('ub.tenantID'::text, true), '0'::text))::bigint"
       } else if (conn.dialect === 'MSSQL2012') {
         dbDefault = "CAST(COALESCE(SESSION_CONTEXT(N'ub.tenantID'), '0') as BigInt)"
-      } else if (conn.dialect === 'SQLite3') { //TODO - fts multitenancy
+      } else if (conn.dialect === 'SQLite3') { // TODO - fts multitenancy
 
-      } else  {
+      } else {
         throw new Error(`DB dialect '${conn.dialect}' is not supported by multitenancy mixin for entity ${entityName}`)
       }
       entityMeta.attributes.push({
@@ -74,14 +75,32 @@ function addMultitenancyMixinAttributes(domainJson, serverConfig) {
         allowNull: false,
         readOnly: true,
         defaultView: false,
-        restriction:  {
+        restriction: {
           I: 'Admin',
           S: 'Admin',
-          replaceBy: "-1",
-          U: "Admin"
+          replaceBy: '-1',
+          U: 'Admin'
         },
         defaultValue: dbDefault
       })
     }
+  }
+}
+
+/**
+ * check blobStore exists in server config for each Document type attribute
+ * @param {object<string, {modelName: string, meta: object, lang: object<string, object>}>} domainJson
+ * @param {object} serverConfig
+ */
+function validateAttributesBlobStore (domainJson, serverConfig) {
+  const bsConfig = serverConfig.application.blobStores || []
+  const bsSet = new Set(bsConfig.map(c => c.name))
+  for (const entityName in domainJson) {
+    const entityMeta = domainJson[entityName].meta
+    entityMeta.attributes.forEach(attr => {
+      if ((attr.dataType === 'Document') && attr.storeName && !bsSet.has(attr.storeName)) {
+        throw new Error(`Entity '${entityName}'. Blob store '${attr.storeName}' used by attribute '${attr.name}' ('storeName' property), but such store is not defined in ubConfig`)
+      }
+    })
   }
 }