+++ packages/adminui-vue/CHANGELOG.md
@@ -6,8 +6,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
- - UAutoField: in case attribute used in `attribute-name` prop is not exists in Repository a developer-friendly
-   exception is throws instead of `can not read XXX of undefined`
 
 ### Changed
 
@@ -16,6 +14,47 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.20.17] - 2021-07-18
+### Added
+ - `USelectEnum`: added `repository` prop, same as `USelectEntity` has
+ - `UCardView`: added support for field slots in `Card view`. Same behavior as in `Table view`
+ - `UTableEntity`: prop `withPagination` for controlling the availability of the table pagination
+ - `UTableEntity`: slot `lastTableRow` to add some content at the end of the table
+
+### Fixed
+ - `USelectEntity`: fixed `More` button for non editable controls - after the click
+   the dropdown list was reopened and the same set of rows was fetched instead of the next one.
+ - `UFormRow`: fixed typo in prop name 'descripion' -> 'description'
+ - `UCodeMirror`: fix updating of the control's value when `valueIsJson=true` and user inputed a number
+
+## [5.20.16] - 2021-07-08
+### Added
+ - `UFormRow`: `attributeName` property - if it is passed, control automatically gets attribute label, required and error
+ - new provided value in Vue instances: `validator`. This value contains useful methods for form validation
+ - new helper `validateWithErrorText` for defining of attribute validation functions with an error text
+ - `validationMixin` for the passing of validation defined locally in some Vue component to nested controls (UFormRow for example)
+ - Validation: added support for `$each` captions in validations.  Example:
+   `crimeCases.$each.11.data.$each.333.val.discardReason:caption` => `crimeCases.data.val.discardReason:caption`
+ - UAutoField: in case attribute used in `attribute-name` prop is not exists in Repository a developer-friendly
+   exception is throws instead of `can not read XXX of undefined`
+ - `USelectMultiple`: new prop `additionalButtons`, for add buttons to dropdown before button 'more'
+ - `USelectEntity`: new prop `allow-dictionary-adding`. If defined and user type text what not match any record -
+   will ask for adding a new record into dictionary. Record parameters must be defined in `build-add-dictionary-config` handler
+
+### Changed
+ - refactored and improved validation of forms. [Documentation](https://git-pub.intecracy.com/unitybase/ubjs/-/blob/master/packages/adminui-vue/utils/Form/validation.md)
+ - use `attributeCaptions` section to define captions for custom attributes or redefine for entity ones
+  instead of computed property with `:caption` suffix
+ - `UFormRow`: display attribute description if exists instead of name when we hover on the label
+ - `u-select-entity` & `u-select-multiple` drop-down is shown after fetch, so in case some old values are in options
+   they will not be shown to the user
+  
+### Removed
+ - `getValidationState` provided value since now it is a method of validator
+
+### Fixed
+- `USelectEntity` prevent click on disable menu items
 
 ## [5.20.15] - 2021-06-14
 ### Added