+++ packages/openid-connect/CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.20.18] - 2021-07-18
+### Added
+ - support for id.gov.ua - see  https://id.gov.ua/downloads/IDInfoProcessingD.pdf
+
+## [5.20.17] - 2021-07-08
 ## [5.20.16] - 2021-06-14
 ### Added
  - `openIDConnect`: id_token saving on auth handshake stage.